Rearranging some stuff

+++ src/rain_sprite.cpp
@@ -2,15 +2,15 @@
 
 #include "rain_sprite.h"
 
-Sprite::Sprite(float w, float h, uint texID) : width(w), height(h), textureID(texID) { }
+Sprite::Sprite(uint w, uint h, uint texID) : width(w), height(h), textureID(texID) { }
 
 Sprite::~Sprite() { }
 
-float Sprite::getWidth() {
+uint Sprite::getWidth() {
 	return width;
 }
 
-float Sprite::getHeight() {
+uint Sprite::getHeight() {
 	return height;
 }
